Fix overflow
uint64_t should be used instead of int

+++ fdbserver/workloads/ClientMetric.actor.cpp
@@ -174,7 +174,7 @@ struct ClientMetricWorkload : TestWorkload {
 	ACTOR Future<uint64_t> writeKeysAndGetLatencyVersion(Database cx,
 	                                                     ClientMetricWorkload* self,
 	                                                     int numKeys,
-	                                                     int previousVS) {
+	                                                     uint64_t previousVS) {
 		state int retry = 0;
 		state int max_retry = 10;
 		state int keysLimit = 1;
